requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
spotipy==2.23.0
wavelink>=2.6.4
supabase==2.0.3
asyncpg>=0.28.0
aiohttp>=3.8.5
aiosqlite>=0.19.0
humanize>=4.7.0
python-dateutil>=2.8.2
pytz>=2023.3
youtube_dl==2021.12.17
PyNaCl==1.5.0
Flask==2.3.3
psutil==5.9.5
requests==2.31.0
asyncio==3.4.3
